dataset: factorize georeferenced aerial dataset handling
