+++ deeposlandia/datasets/aerial.py
@@ -9,16 +9,18 @@ from multiprocessing import Pool
 import os
 
 import daiquiri
+import numpy as np
+from osgeo import gdal
 from PIL import Image
 
-from deeposlandia.datasets import Dataset
-from deeposlandia import utils
+from deeposlandia.datasets import GeoreferencedDataset
+from deeposlandia import geometries, utils
 
 
 logger = daiquiri.getLogger(__name__)
 
 
-class AerialDataset(Dataset):
+class AerialDataset(GeoreferencedDataset):
     """Dataset structure inspired from AerialImageDataset, a dataset released
     by Inria
 
@@ -40,9 +42,9 @@ class AerialDataset(Dataset):
         Dataset which is completed by a glossary file that describes the
         dataset labels and images
         """
-        self.tile_size = tile_size
-        img_size = utils.get_image_size_from_tile(self.tile_size)
-        super().__init__(img_size)
+        # self.tile_size = tile_size
+        # img_size = utils.get_image_size_from_tile(self.tile_size)
+        super().__init__(tile_size)
         self.add_label(
             label_id=0, label_name="background", color=0, is_evaluate=True
         )
@@ -50,14 +52,15 @@ class AerialDataset(Dataset):
             label_id=1, label_name="building", color=255, is_evaluate=True
         )
 
-    def _preprocess(self, image_filename, output_dir, labelling):
+
+    def _preprocess_for_training(self, image_filename, output_dir, nb_images):
         """Resize/crop then save the training & label images
 
         Parameters
         ----------
         image_filename : str
             Full path towards the image on the disk
-        datadir : str
+        output_dir : str
             Output path where preprocessed image must be saved
 
         Returns
@@ -65,128 +68,119 @@ class AerialDataset(Dataset):
         dict
             Key/values with the filenames and label ids
         """
-        img_in = Image.open(image_filename)
-        raw_img_size = img_in.size[0]
+        raster = gdal.Open(image_filename)
+        raw_img_width = raster.RasterXSize
+        raw_img_height = raster.RasterYSize
+        image_data = raster.ReadAsArray()
+        image_data = np.swapaxes(image_data, 0, 2)
         result_dicts = []
-        # crop tile_size*tile_size tiles into 5000*5000 raw images
-        for x in range(0, raw_img_size, self.tile_size):
-            for y in range(0, raw_img_size, self.tile_size):
-                tile = img_in.crop(
-                    (x, y, x + self.tile_size, y + self.tile_size)
-                )
-                tile = utils.resize_image(tile, self.image_size)
-                img_id = int(
-                    (
-                        raw_img_size / self.tile_size * x / self.tile_size
-                        + y / self.tile_size
-                    )
-                )
-                basename_decomp = os.path.splitext(
-                    os.path.basename(image_filename)
-                )
-                new_in_filename = (
-                    basename_decomp[0] + "_" + str(img_id) + basename_decomp[1]
-                )
-                new_in_path = os.path.join(
-                    output_dir, "images", new_in_filename
-                )
-                tile.save(new_in_path.replace(".tif", ".png"))
-                result_dicts.append(
-                    {
-                        "raw_filename": image_filename,
-                        "image_filename": new_in_path,
-                    }
-                )
+        logger.info("Image filename: %s", image_filename)
+        logger.info("Raw image size: %s, %s", raw_img_width, raw_img_height)
 
-        if labelling:
-            label_filename = image_filename.replace("images/", "gt/")
-            img_out = Image.open(label_filename)
-            for x in range(0, raw_img_size, self.tile_size):
-                for y in range(0, raw_img_size, self.tile_size):
-                    tile = img_out.crop(
-                        (x, y, x + self.tile_size, y + self.tile_size)
-                    )
-                    tile = utils.resize_image(tile, self.image_size)
-                    img_id = int(
-                        (
-                            raw_img_size / self.tile_size * x / self.tile_size
-                            + y / self.tile_size
-                        )
-                    )
-                    basename_decomp = os.path.splitext(
-                        os.path.basename(image_filename)
-                    )
-                    new_out_filename = (
-                        basename_decomp[0]
-                        + "_"
-                        + str(img_id)
-                        + basename_decomp[1]
-                    )
-                    new_out_path = os.path.join(
-                        output_dir, "labels", new_out_filename
-                    )
-                    tile.save(new_out_path.replace(".tif", ".png"))
-                    labels = utils.build_labels(
-                        tile, self.label_ids, dataset="aerial"
-                    )
-                    result_dicts[img_id]["label_filename"] = new_out_path
-                    result_dicts[img_id]["labels"] = labels
+        label_filename = image_filename.replace("images", "gt")
+        label_raster = gdal.Open(label_filename)
+        labels = label_raster.ReadAsArray()
 
+        nb_attempts = 0
+        image_counter = 0
+        empty_image_counter = 0
+        while image_counter < nb_images and nb_attempts < 2 * nb_images:
+            # randomly pick an image
+            x = np.random.randint(0, raw_img_width - self.image_size)
+            y = np.random.randint(0, raw_img_height - self.image_size)
+
+            tile_data = image_data[
+                x:(x + self.image_size), y:(y + self.image_size)
+            ]
+            tile_image = Image.fromarray(tile_data)
+            mask = labels[
+                x:(x + self.image_size), y:(y + self.image_size)
+            ]
+            label_dict = utils.build_labels(
+                mask, range(self.get_nb_labels()), "aerial"
+            )
+            labelled_image = utils.build_image_from_config(mask, self.labels)
+            if np.unique(mask).shape[0] > 1:
+                tiled_results = self._serialize(
+                    tile_image,
+                    labelled_image,
+                    label_dict,
+                    image_filename,
+                    output_dir,
+                    x,
+                    y,
+                    "nw",
+                )
+                result_dicts.append(tiled_results)
+                image_counter += 1
+                tile_image_ne = tile_image.transpose(Image.FLIP_LEFT_RIGHT)
+                labelled_image_ne = labelled_image.transpose(
+                    Image.FLIP_LEFT_RIGHT
+                )
+                tiled_results_ne = self._serialize(
+                    tile_image_ne,
+                    labelled_image_ne,
+                    label_dict,
+                    image_filename,
+                    output_dir,
+                    x,
+                    y,
+                    "ne",
+                )
+                result_dicts.append(tiled_results_ne)
+                image_counter += 1
+                tile_image_sw = tile_image.transpose(Image.FLIP_TOP_BOTTOM)
+                labelled_image_sw = labelled_image.transpose(
+                    Image.FLIP_TOP_BOTTOM
+                )
+                tiled_results_sw = self._serialize(
+                    tile_image_sw,
+                    labelled_image_sw,
+                    label_dict,
+                    image_filename,
+                    output_dir,
+                    x,
+                    y,
+                    "sw",
+                )
+                result_dicts.append(tiled_results_sw)
+                image_counter += 1
+                tile_image_se = tile_image_sw.transpose(Image.FLIP_LEFT_RIGHT)
+                labelled_image_se = labelled_image_sw.transpose(
+                    Image.FLIP_LEFT_RIGHT
+                )
+                tiled_results_se = self._serialize(
+                    tile_image_se,
+                    labelled_image_se,
+                    label_dict,
+                    image_filename,
+                    output_dir,
+                    x,
+                    y,
+                    "se",
+                )
+                result_dicts.append(tiled_results_se)
+                image_counter += 1
+                del tile_image_se, tile_image_sw, tile_image_ne
+                del labelled_image_se, labelled_image_sw, labelled_image_ne
+            else:
+                if empty_image_counter < 0.1 * nb_images:
+                    tiled_results = self._serialize(
+                        tile_image,
+                        labelled_image,
+                        label_dict,
+                        image_filename,
+                        output_dir,
+                        x,
+                        y,
+                        "nw",
+                    )
+                    result_dicts.append(tiled_results)
+                    image_counter += 1
+                    empty_image_counter += 1
+            nb_attempts += 1
+        del raster
+        logger.info(
+            "Generate %s images after %s attempts.", image_counter, nb_attempts
+        )
         return result_dicts
-
-    def populate(
-        self,
-        output_dir,
-        input_dir,
-        nb_images=None,
-        aggregate=False,
-        labelling=True,
-        nb_processes=1,
-    ):
-        """ Populate the dataset with images contained into `datadir` directory
-
-        Parameters
-        ----------
-        output_dir : str
-            Path of the directory where the preprocessed image must be saved
-        input_dir : str
-            Path of the directory that contains input images
-        nb_images : integer
-            Number of images to be considered in the dataset; if None, consider
-        the whole repository
-        aggregate : bool
-            Label aggregation parameter, useless for this dataset, but kept for
-        class method genericity
-        labelling : boolean
-            If True labels are recovered from dataset, otherwise dummy label
-        are generated
-        nb_processes : int
-            Number of processes on which to run the preprocessing
-        """
-        image_list = os.listdir(os.path.join(input_dir, "images"))
-        image_list_longname = [
-            os.path.join(input_dir, "images", l)
-            for l in image_list
-            if not l.startswith(".")
-        ][:nb_images]
-        logger.info(
-            "Getting %s images to preprocess...", len(image_list_longname)
-        )
-        if nb_processes == 1:
-            for x in image_list_longname:
-                self.image_info.append(
-                    self._preprocess(x, output_dir, labelling)
-                )
-        else:
-            with Pool(processes=nb_processes) as p:
-                self.image_info = p.starmap(
-                    self._preprocess,
-                    [(x, output_dir, labelling) for x in image_list_longname],
-                )
-        self.image_info = [
-            item for sublist in self.image_info for item in sublist
-        ]
-        logger.info(
-            "Saved %s images in the preprocessed dataset.",
-            len(self.image_info),
-        )
